Little refactor, braking code into multiple files, and adding authentication
